Switch to a simplified build system
We don't use any of the node based systems right now (or ever?)
so there is no reason to complicate the build process with it.

Previously, contributors have had problems using this system so
hopefully this change lowers that hurdle.

Furthermore, I have update the README to reflect this change so
hopefully that also makes things easier.