Revert "Fixup github plugin to return correct URL"
This reverts commit ca80a7c3662e47a6392b6f1068459e0a83066402.

You need this fix for it to work locally, but readthedocs is doing
something horrible... well just in general with how they change the
site, but also such that this ends up duplicating the source root.

But when I move us away from RTD in a month or two (hopefully relatively
soon) chances are I'm going to need to revert this revert. YAY!